% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_fit.r
\name{.lm_fit}
\alias{.lm_fit}
\title{Linear Model Fitter}
\usage{
.lm_fit(x, y, intercept = FALSE)
}
\arguments{
\item{x}{The input data matrix.}

\item{y}{The vector/matrix of independent variable(s).}

\item{intercept}{Should an intercept term be included?}
}
\description{
A basic linear model fitter.
}
\examples{
\dontrun{
library(linmod2)

n <- 10
p <- 3
x <- matrix(rnorm(n*p), n, p)
y <- rnorm(n)

.lm_fit(x, y)
}

}
